+++ src/mesh/v4.rs
@@ -1,8 +1,8 @@
 use binrw::BinReaderExt;
 
+use super::v2::DEFAULT_VERTEX_TANGENT;
 use super::v2::{Face2, Vertex2};
 use super::v3::Lod3;
-use super::DEFAULT_VERTEX_TANGENT;
 
 #[binrw::binrw]
 #[brw(little)]
